telegram-alert/requirements.txt
google-cloud-bigquery
pandas==1.1.2